refactor: move AspectRatio into SharedComponents
