+++ src/components/Posts/PostList/PostListItem/PostListItem.tsx
@@ -7,12 +7,12 @@ import LineClamp from "../../../LineClamp";
 import OakLink from "../../../OakLink";
 import BoxBorders from "../../../SpriteSheet/BrushSvgs/BoxBorders";
 import { P, Heading, HeadingTag } from "../../../Typography";
-import AspectRatio from "../../../AspectRatio";
 import { ResolveOakHrefProps } from "../../../../common-lib/urls";
 import formatDate from "../../../../utils/formatDate";
 
 import PostListItemImage from "./PostListItemImage";
 
+import AspectRatio from "@/components/SharedComponents/AspectRatio";
 import Flex from "@/components/SharedComponents/Flex";
 import Box from "@/components/SharedComponents/Box";
