+++ src/components/Posts/UpcomingWebinarWall/UpcomingWebinarWall.stories.tsx
@@ -1,9 +1,8 @@
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 
-import AspectRatio from "../../AspectRatio";
-
 import Component from ".";
 
+import AspectRatio from "@/components/SharedComponents/AspectRatio";
 import Box from "@/components/SharedComponents/Box";
 
 export default {
