+++ src/components/Posts/WebinarVideo/WebinarVideo.tsx
@@ -1,5 +1,4 @@
 import { SerializedWebinar } from "../../../pages/webinars/[webinarSlug]";
-import AspectRatio from "../../AspectRatio";
 import CMSVideo from "../../CMSVideo";
 import BoxBorders from "../../SpriteSheet/BrushSvgs/BoxBorders";
 import WebinarRegistration, {
@@ -9,6 +8,8 @@ import isUpcomingWebinar from "../../../utils/isUpcomingWebinar";
 import UpcomingWebinarWall, {
   useUpcomingWebinarWall,
 } from "../UpcomingWebinarWall";
+
+import AspectRatio from "@/components/SharedComponents/AspectRatio";
 
 type WebinarVideoProps = {
   webinar: SerializedWebinar;
