+++ src/components/SharedComponents/AspectRatio/AspectRatio.stories.tsx
